Refactor UserPrefs into package
