+++ src/main/java/seedu/address/model/ModelManager.java
@@ -16,6 +16,7 @@ import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
 import seedu.address.model.person.exceptions.PersonNotFoundException;
 import seedu.address.model.tag.Tag;
+import seedu.address.model.user.UserPrefs;
 
 /**
  * Represents the in-memory model of the address book data.
