+++ src/test/java/seedu/address/logic/commands/SearchCommandTest.java
@@ -18,7 +18,7 @@ import seedu.address.logic.UndoRedoStack;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
-import seedu.address.model.UserPrefs;
+import seedu.address.model.user.UserPrefs;
 import seedu.address.model.person.DetailsContainKeyphrasePredicate;
 import seedu.address.model.person.ReadOnlyPerson;
