+++ src/test/java/seedu/address/logic/commands/SelectCommandTest.java
@@ -22,7 +22,7 @@ import seedu.address.logic.UndoRedoStack;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
-import seedu.address.model.UserPrefs;
+import seedu.address.model.user.UserPrefs;
 import seedu.address.ui.testutil.EventsCollectorRule;
 
 /**
