+++ src/test/java/seedu/address/model/ModelManagerTest.java
@@ -13,6 +13,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.user.UserPrefs;
 import seedu.address.testutil.AddressBookBuilder;
 
 public class ModelManagerTest {
