+++ src/test/java/seedu/address/storage/JsonUserPrefsStorageTest.java
@@ -14,7 +14,7 @@ import org.junit.rules.TemporaryFolder;
 
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.FileUtil;
-import seedu.address.model.UserPrefs;
+import seedu.address.model.user.UserPrefs;
 
 public class JsonUserPrefsStorageTest {
