+++ src/test/java/seedu/address/storage/StorageManagerTest.java
@@ -16,7 +16,7 @@ import seedu.address.commons.events.model.AddressBookChangedEvent;
 import seedu.address.commons.events.storage.DataSavingExceptionEvent;
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.UserPrefs;
+import seedu.address.model.user.UserPrefs;
 import seedu.address.ui.testutil.EventsCollectorRule;
 
 public class StorageManagerTest {
